+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- `99designs/gqlgen` trace instrumentation
+
 ## [0.20.0] - 2021-04-23
 
 ### Changed